Load spaces from hub
