+++ src/helpers/skins.ts
@@ -1,0 +1,7 @@
+const requireFile = require.context(
+  '@bonustrack/snapshot-spaces/spaces/',
+  true,
+  /[\w-]+\.scss$/
+);
+
+requireFile.keys().map(file => requireFile(file));
