+++ src/main.ts
@@ -11,6 +11,7 @@ import store from '@/store';
 import mixins from '@/mixins';
 import i18n from '@/i18n';
 import '@/auth';
+import '@/helpers/skins';
 import '@/style.scss';
 
 Vue.use(infiniteScroll);
